Move connection to dao package
