+++ src/main/java/by/epamtc/digapply/controller/Controller.java
@@ -3,8 +3,8 @@ package by.epamtc.digapply.controller;
 import by.epamtc.digapply.command.Command;
 import by.epamtc.digapply.command.CommandResult;
 import by.epamtc.digapply.command.factory.CommandFactory;
-import by.epamtc.digapply.connection.ConnectionPool;
-import by.epamtc.digapply.connection.ConnectionPoolException;
+import by.epamtc.digapply.dao.connection.ConnectionPool;
+import by.epamtc.digapply.dao.connection.ConnectionPoolException;
 import by.epamtc.digapply.command.PagePath;
 import by.epamtc.digapply.command.RequestParameter;
 import by.epamtc.digapply.service.ServiceException;
