+++ src/main/java/by/epamtc/digapply/dao/JdbcOperator.java
@@ -1,7 +1,7 @@
 package by.epamtc.digapply.dao;
 
-import by.epamtc.digapply.connection.ConnectionPool;
-import by.epamtc.digapply.connection.ConnectionPoolException;
+import by.epamtc.digapply.dao.connection.ConnectionPool;
+import by.epamtc.digapply.dao.connection.ConnectionPoolException;
 import by.epamtc.digapply.entity.Identifiable;
 import by.epamtc.digapply.dao.mapper.RowMapper;
 import org.apache.logging.log4j.LogManager;
